remove geom2d_utils
