+++ src/prbench/envs/geom2d/stickbutton2d.py
@@ -15,9 +15,9 @@ from numpy.typing import NDArray
 from relational_structs import Object, ObjectCentricState
 from relational_structs.utils import create_state_from_dict
 
-from prbench.envs.geom2d.geom2d_utils import ConstantObjectGeom2DEnv
 from prbench.envs.geom2d.utils import (
     BLACK,
+    ConstantObjectGeom2DEnv,
     CRVRobotActionSpace,
     SE2Pose,
     create_walls_from_world_boundaries,
